添加统一异常处理、统一返回报文组装、统一返回码定义

+++ defender/defender-dao/src/main/java/com/cib/defender/dao/mapper/RoleMapper.java
@@ -3,13 +3,13 @@ package com.cib.defender.dao.mapper;
 import com.cib.defender.dao.entity.Role;
 
 public interface RoleMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Role record);
 
     int insertSelective(Role record);
 
-    Role selectByPrimaryKey(Integer id);
+    Role selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(Role record);
